Combine files to redis_operations.py
