+++ scripts/redis_operations.py
@@ -1,3 +1,5 @@
+import scrape_lines
+import scrape_line
 import redis
 
 
@@ -34,7 +36,13 @@ stop_example = {
 """
 
 
-def connect(database):
+# def main():
+#     SELECTED_LINES = ['01N', '02K', '03K']
+#     stops = get_line_stops(SELECTED_LINES)
+#     print(stops)
+
+
+def connect(database=0):
     """Creates the connection object to the database and prints basic information to the console.
 
     Arguments:
@@ -45,8 +53,11 @@ def connect(database):
     """
 
     r = redis.Redis(host='localhost', port=6379, db=database)
-    # print(r.info())
     return r
+
+# ---------------------------------------------------------------------------- #
+#                                    SAVING                                    #
+# ---------------------------------------------------------------------------- #
 
 
 def save_stops(r, stops):
@@ -107,3 +118,60 @@ def save(stops=None, lines=None):
         save_lines(r, lines)
     if stops is not None:
         save_stops(r, stops)
+
+
+# ---------------------------------------------------------------------------- #
+#                                    LOADING                                   #
+# ---------------------------------------------------------------------------- #
+
+
+def get_line_stops(selected_lines, db=0):
+
+    selected_stops = set()
+
+    # Initialize Redis client object
+    r = redis.Redis(host='localhost', port=6379, db=db)
+
+    # Load lines from Redis cache into dictionary
+    selected_lines_uids = set([int(l) for l in r.hmget('lines', selected_lines)])
+    # print(selected_lines_uids)
+
+    for line_uid in selected_lines_uids:
+        for direction in (1, 2):
+            lsuid = 'line{}:stops:direction{}'.format(line_uid, direction)
+            stops = [int(s) for s in r.hgetall(lsuid).keys()]
+            selected_stops.update(stops)
+
+    return selected_stops
+
+
+# ---------------------------------------------------------------------------- #
+#                             UPDATE INFRASTRUCTURE                            #
+# ---------------------------------------------------------------------------- #
+
+
+def redis_update_infrastructure(db=0):
+
+    # Initialize Redis client object
+    r = redis.Redis(host='localhost', port=6379, db=db)
+
+    # Clean all lines in db
+    r.flushdb()
+
+    # Scrape all indivudual lines and save them
+    # TODO: Do not use saving by default inside scrape_lines
+    scrape_lines.scrape_lines()
+
+    # Load all line UIDs in memory
+    lines = [l.decode('utf-8') for l in r.hvals('lines')]
+
+    # Use line UID to scrape individual lines
+    for line in lines:
+        scrape_line.scrape_line(line)
+
+    # Save database to memory
+    r.save()
+
+
+# if __name__ == "__main__":
+#     main()
